@@ -23,35 +24,35 @@
   },
   "dependencies": {
     "@abraham/reflection": "^0.12.0",
-    "@clerk/nextjs": "^4.29.6",
-    "@fortawesome/fontawesome-svg-core": "^6.5.1",
-    "@fortawesome/free-brands-svg-icons": "^6.5.1",
-    "@fortawesome/free-regular-svg-icons": "^6.5.1",
-    "@fortawesome/free-solid-svg-icons": "^6.5.1",
+    "@clerk/nextjs": "^5.3.7",
+    "@fortawesome/fontawesome-svg-core": "^6.6.0",
+    "@fortawesome/free-brands-svg-icons": "^6.6.0",
+    "@fortawesome/free-regular-svg-icons": "^6.6.0",
+    "@fortawesome/free-solid-svg-icons": "^6.6.0",
     "@fullpage/react-fullpage": "^0.1.42",
     "@linaria/babel-preset": "^5.0.4",
-    "@linaria/core": "^6.0.0",
-    "@linaria/react": "^6.0.0",
+    "@linaria/core": "^6.2.0",
+    "@linaria/react": "^6.2.1",
     "@linaria/shaker": "^5.0.3",
     "@linaria/webpack-loader": "^5.0.4",
     "@popperjs/core": "^2.11.8",
-    "@prisma/client": "^5.9.1",
-    "@reduxjs/toolkit": "^2.1.0",
-    "@vercel/analytics": "^1.1.3",
-    "axios": "^1.6.7",
-    "clsx": "^2.1.0",
+    "@prisma/client": "^5.19.0",
+    "@reduxjs/toolkit": "^2.2.7",
+    "@vercel/analytics": "^1.3.1",
+    "axios": "^1.7.6",
+    "clsx": "^2.1.1",
     "comlink": "^4.4.1",
-    "compression-webpack-plugin": "^11.0.0",
+    "compression-webpack-plugin": "^11.1.0",
     "cross-env": "^7.0.3",
-    "flowbite": "^2.2.1",
-    "flowbite-datepicker": "^1.2.6",
-    "flowbite-react": "^0.7.2",
-    "formik": "^2.4.5",
+    "flowbite": "^2.5.1",
+    "flowbite-datepicker": "^1.3.0",
+    "flowbite-react": "^0.10.1",
+    "formik": "^2.4.6",
     "inversify-esm": "^2.0.2",
     "inversify-hooks-esm": "^2.1.2",
     "inversify-props-esm": "^2.0.2",
-    "jose": "^5.2.1",
-    "next": "14.1.0",
+    "jose": "^5.8.0",
+    "next": "14.2.7",
     "next-compose-plugins": "^2.2.1",
     "next-linaria": "^1.0.1-beta",
     "next-optimized-images": "^2.6.2",
@@ -59,43 +60,43 @@
     "next-redux-cookie-wrapper": "^2.2.1",
     "next-redux-wrapper": "^8.1.0",
     "purge-fontawesome": "^1.1.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "react-dropzone": "^14.2.3",
     "react-popper": "^2.3.0",
-    "react-redux": "^9.1.0",
-    "react-toastify": "^10.0.4",
-    "read-excel-file": "^5.7.1",
-    "sass": "^1.70.0",
-    "tailwindcss": "^3.4.1",
+    "react-redux": "^9.1.2",
+    "react-toastify": "^10.0.5",
+    "read-excel-file": "^5.8.5",
+    "sass": "^1.77.8",
+    "tailwindcss": "^3.4.10",
     "ts-node": "^10.9.2",
     "windicss": "^3.5.6",
     "windicss-webpack-plugin": "^1.8.0",
-    "yup": "^1.3.3"
+    "yup": "^1.4.0"
   },
   "devDependencies": {
-    "@commitlint/cli": "^18.6.0",
-    "@commitlint/config-conventional": "^18.6.0",
+    "@commitlint/cli": "^19.4.1",
+    "@commitlint/config-conventional": "^19.4.1",
     "@types/lodash-es": "^4.17.12",
-    "@types/node": "^20.11.16",
-    "@types/react": "^18.2.54",
-    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@types/node": "^22.5.1",
+    "@types/react": "^18.3.5",
+    "@typescript-eslint/eslint-plugin": "^8.3.0",
     "add": "^2.0.6",
-    "autoprefixer": "^10.4.17",
-    "eslint": "8.56.0",
-    "eslint-config-next": "14.1.0",
+    "autoprefixer": "^10.4.20",
+    "eslint": "8.57.0",
+    "eslint-config-next": "14.2.7",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "husky": "^9.0.10",
-    "lint-staged": "^15.2.2",
+    "eslint-plugin-prettier": "^5.2.1",
+    "husky": "^9.1.5",
+    "lint-staged": "^15.2.9",
     "lodash-es": "^4.17.21",
-    "postcss": "^8.4.34",
-    "postcss-import": "^16.0.0",
-    "prettier": "^3.2.5",
-    "prisma": "^5.9.1",
-    "typescript": "^5.3.3"
+    "postcss": "^8.4.41",
+    "postcss-import": "^16.1.0",
+    "prettier": "^3.3.3",
+    "prisma": "^5.19.0",
+    "typescript": "^5.5.4"
   },
-  "packageManager": "yarn@3.6.4",
+  "packageManager": "yarn@4.4.1",
   "lint-staged": {
     "**/*": "prettier --write --ignore-unknown"
   },
