@@ -1,14 +1,17 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
-const { resolve, join } = require('path')
-const glob = require('glob')
+import { resolve, join, dirname } from 'path'
+import glob from 'glob'
+import { fileURLToPath } from 'url'
 
-const withPlugins = require('next-compose-plugins')
-const withPWA = require('next-pwa')
-const optimizedImages = require('next-optimized-images')
+import withPlugins from 'next-compose-plugins'
+import withPWA from 'next-pwa'
+import optimizedImages from 'next-optimized-images'
 // const WindiCSSWebpackPlugin = require('windicss-webpack-plugin')
-const CompressionWebpackPlugin = require('compression-webpack-plugin')
-const PurgeFontawesomePlugin = require('purge-fontawesome/webpack-plugin')
+import CompressionWebpackPlugin from 'compression-webpack-plugin'
+import PurgeFontawesomePlugin from 'purge-fontawesome/webpack-plugin.js'
 
+const __filename = fileURLToPath(import.meta.url) // get the resolved path to the file
+const __dirname = dirname(__filename) // get the name of the directory
 const isProduction = process.env.NODE_ENV === 'production'
 const isNextExportCommand = process.env.NEXT_EXPORT_COMMAND === 'true'
 
@@ -19,26 +22,36 @@ const nextConfigurations = {
   images: {
     remotePatterns: [
       {
+        protocol: 'https',
+        port: '',
         hostname: 'floatui.com'
       },
       {
+        protocol: 'https',
+        port: '',
         hostname: '**.clerk.dev'
       },
       {
+        protocol: 'https',
+        port: '',
         hostname: '**.clerk.com'
       }
     ],
-    ...(isNextExportCommand ? {
-      loader: 'akamai',
-      unoptimized: true,
-      path: '/'
-    } : {})
+    ...(isNextExportCommand
+      ? {
+          loader: 'akamai',
+          unoptimized: true,
+          path: '/'
+        }
+      : {})
   },
-  ...(isNextExportCommand ? {
-    exportPathMap: async () => ({
-      '/': { page: '/' }
-    })
-  } : {}),
+  ...(isNextExportCommand
+    ? {
+        exportPathMap: async () => ({
+          '/': { page: '/' }
+        })
+      }
+    : {}),
   webpack(config) {
     config.resolve.alias['@$'] = resolve(__dirname, 'src')
     config.resolve.alias['linaria/loader'] = '@linaria/webpack-loader'
@@ -50,27 +63,27 @@ const nextConfigurations = {
       )
       config.plugins.push(
         new PurgeFontawesomePlugin({
-          paths: [
-            glob.sync(join(__dirname, 'src/**/*'), { nodir: true })
-          ]
+          paths: [glob.sync(join(__dirname, 'src/**/*'), { nodir: true })]
         })
       )
     }
     return config
   },
-  ...(isProduction ? {
-    eslint: {
-      ignoreDuringBuilds: true,
-    },
-    pwa: {
-      dest: 'public',
-      maximumFileSizeToCacheInBytes: 8 * 1024 * 1024
-    },
-    swcMinify: true,
-  } : {})
+  ...(isProduction
+    ? {
+        eslint: {
+          ignoreDuringBuilds: true
+        },
+        pwa: {
+          dest: 'public',
+          maximumFileSizeToCacheInBytes: 8 * 1024 * 1024
+        },
+        swcMinify: true
+      }
+    : {})
 }
 
-module.exports = withPlugins([
-  ...(isNextExportCommand ? [optimizedImages] : []),
-  ...(isProduction ? [withPWA] : [])
-], nextConfigurations)
+export default withPlugins(
+  [...(isNextExportCommand ? [optimizedImages] : []), ...(isProduction ? [withPWA] : [])],
+  nextConfigurations
+)
