@@ -1,4 +1,4 @@
-module.exports = {
+const config = {
   extends: ['@commitlint/config-conventional'],
   rules: {
     'body-max-line-length': [0, 'always', Infinity],
@@ -10,3 +10,5 @@ module.exports = {
     },
   },
 }
+
+export default config;
