@@ -1,5 +1,5 @@
 import { forwardRef } from 'react'
-import type { Ref, ComponentType, ForwardRefRenderFunction } from 'react'
+import type { Ref, ComponentType } from 'react'
 
 export interface PropsWithInnerRef<RefElement = HTMLDivElement> {
   innerRef?: Ref<RefElement>
@@ -9,8 +9,6 @@ export function addRefProps<
   RefElement = HTMLDivElement,
   ComponentProps extends PropsWithInnerRef<RefElement> = PropsWithInnerRef<RefElement>
 >(Component: ComponentType<ComponentProps>) {
-  const forwardRefCallback: ForwardRefRenderFunction<RefElement, ComponentProps> = (props, ref) => (
-    <Component innerRef={ref} {...props} />
-  )
+  const forwardRefCallback = (props, ref) => <Component innerRef={ref} {...props} />
   return forwardRef<RefElement, ComponentProps>(forwardRefCallback)
 }
